Move open back to SChunk module
